mimo: fold bootstrap into the runtime module
Merge the per-rank assembly (MimoRuntime, mimo_model_provider, build_mimo_runtime,
the role/seed helpers and the p2p-communicator builder) into runtime.py, which
already held the DDP-wrap helpers, and drop bootstrap.py. This removes the
bootstrap<->runtime import edge and keeps all per-rank runtime setup in one place.
